Support Tomee JAX-WS (OpenEJB/CXF)
