+++ instrumentation/jaxws/jaxws-2.0-cxf-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/cxf/CxfInstrumentationModule.java
@@ -19,6 +19,6 @@ public class CxfInstrumentationModule extends InstrumentationModule {
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return Collections.singletonList(new EndpointImplTypeInstrumentation());
+    return Collections.singletonList(new JaxWsServerFactoryBeanInstrumentation());
   }
 }
