+++ instrumentation/jaxws/jaxws-2.0-cxf-3.0/library/src/main/java/io/opentelemetry/javaagent/instrumentation/cxf/TracingStartInInterceptor.java
@@ -11,7 +11,7 @@ import org.apache.cxf.message.Message;
 import org.apache.cxf.phase.AbstractPhaseInterceptor;
 import org.apache.cxf.phase.Phase;
 
-public class TracingStartInInterceptor extends AbstractPhaseInterceptor {
+public class TracingStartInInterceptor extends AbstractPhaseInterceptor<Message> {
 
   public TracingStartInInterceptor() {
     super(Phase.PRE_INVOKE);
